Move DeleteData functions to platform-specific files
